01-intro-to-ai-agents/code_samples/requirements.txt
agent-framework-core